Completed All Ordering Scenes and Back Button
Delivery, Take-Out, and Dine-In scenes are completed.

Order related information are currently printed out in the console, they will be stored in the database when it's set up.

A back button along with a prefab has also been created. Drag and drop the under Canvas to use the back button. Make sure Canvas size is correct, no additional adjustments will be required